backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic
qrcode
pillow
pyzbar
numpy
onnxruntime
opencv-python-headless
prometheus-client
loguru
fastapi-sentry-sdk
reportlab
python-dotenv
